# backend-filemanagement/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/billings/store": {
            "post": {
                "tags": [
                    "Billings"
                ],
                "summary": "Create a new billing record",
                "description": "Creates a new billing record for the authenticated user",
                "operationId": "storeBilling",
                "requestBody": {
                    "description": "Billing data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Billing record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Billing record created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/billings/show/{id}": {
            "get": {
                "tags": [
                    "Billings"
                ],
                "summary": "Show billing record by ID",
                "description": "Returns a billing record by its ID for the authenticated user",
                "operationId": "showBilling",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the billing record to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Billing record not found"
                    }
                }
            }
        },
        "api/billings/update/{id}": {
            "put": {
                "tags": [
                    "Billings"
                ],
                "summary": "Update billing record by ID",
                "description": "Updates an existing billing record by its ID for the authenticated user",
                "operationId": "updateBilling",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the billing record to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Billing data to update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Billing record updated successfully"
                    },
                    "404": {
                        "description": "Billing record not found"
                    }
                }
            }
        },
...
    }
}